[UPDATE] final SaleDetail

+++ src/main/java/com/rmit/trading_backend/controller/ProductController.java
@@ -111,7 +111,7 @@ public class ProductController {
 
     // DELETE ONE PRODUCT
     @DeleteMapping("/products/{id}")
-    public ResponseEntity<String> deleteProduct(@PathVariable("id") int id) {
+    public ResponseEntity<String> deleteProduct(@PathVariable("id") long id) {
         try {
             if (productRepository.findById(id).isPresent()) {
                 productRepository.deleteById(id);
@@ -139,7 +139,7 @@ public class ProductController {
 
     // UPDATE
     @PutMapping("products/{id}")
-    public ResponseEntity<Product> updateCustomerById(@PathVariable("id") int id, @RequestBody Product product) {
+    public ResponseEntity<Product> updateCustomerById(@PathVariable("id") long id, @RequestBody Product product) {
         Optional<Product> updatedProduct = productRepository.findById(id);
         try {
             if (updatedProduct.isPresent()) {
